feat(discord): polish mobile App Launcher sheet and consolidate types
Add half/full snap with drag handle, overflow View More, dynamic promo
titles, and fold launcher/select-menu helpers into shared constants.
