+++ app/utils/constants.ts
@@ -1,5 +1,50 @@
 import type { InjectionKey, Ref } from "vue";
-import type { StringSelectMenuOption } from "~/types/string-select-menu";
+import type {
+	HomeAccent,
+	HomeDashboardMember,
+	HomeFeature,
+	HomeStat,
+	HomeTestimonial,
+	LoggingEventDetail,
+	ModerationAction,
+	Profile,
+	ShowcaseCommand,
+	SlashCommandApp,
+	SlashCommandAppName,
+} from "~/types/constants";
+import type {
+	DiscordAppLauncherEntry,
+	DiscordAppLauncherListView,
+	DiscordAppLauncherPromo,
+	DiscordAppLauncherSheetSnap,
+	ResolveDiscordAppLauncherSheetSnapOptions,
+	StringSelectMenuPlacement,
+} from "~/types/discord";
+import { Colors } from "~/types/constants";
+
+export type {
+	HomeAccent,
+	HomeDashboardMember,
+	HomeFeature,
+	HomeStat,
+	HomeTestimonial,
+	LoggingEventDetail,
+	ModerationAction,
+	OtherApp,
+	Profile,
+	ProfileName,
+	ShowcaseCommand,
+	SlashCommandApp,
+	SlashCommandAppName,
+	UIColors,
+} from "~/types/constants";
+export { BrandingColors, Colors } from "~/types/constants";
+export type {
+	DiscordAppLauncherSheetSnap,
+	ResolveDiscordAppLauncherSheetSnapOptions,
+	StringSelectMenuOption,
+	StringSelectMenuPlacement,
+} from "~/types/discord";
 
 export const robotBlockingPageProps =
 	"nosnippet,notranslate,noimageindex,noarchive,max-snippet:-1,max-image-preview:none,max-video-preview:-1";
@@ -14,97 +59,6 @@ export const colors = [
 	"accent",
 	"neutral",
 ] as const;
-
-export type UIColors = (typeof colors)[number];
-
-export type HomeAccent =
-	| "primary"
-	| "secondary"
-	| "success"
-	| "error"
-	| "info"
-	| "warning"
-	| "accent"
-	| "spectrum-red"
-	| "spectrum-orange"
-	| "spectrum-yellow"
-	| "spectrum-green"
-	| "spectrum-blue"
-	| "spectrum-purple";
-
-export enum BrandingColors {
-	Secondary = "#fd171b",
-}
-
-export enum Colors {
-	Amber = "#ffc107",
-	Amber300 = "#ffd54f",
-	DeepOrange = "#ff5722",
-	LightBlue = "#03a9f4",
-	Orange = "#ff9800",
-	Red = "#f44336",
-	Red300 = "#e57373",
-	Yellow = "#ffeb3b",
-	Yellow300 = "#fff176",
-}
-
-export interface HomeStat {
-	accent: HomeAccent;
-	label: string;
-	value: string;
-}
-
-export interface HomeFeature {
-	accent?: HomeAccent;
-	big: boolean;
-	description: string;
-	icon: string;
-	title: string;
-}
-
-export interface HomeDashboardMember {
-	avatarClass: string;
-	badgeClass: string;
-	initials: string;
-	name: string;
-	nameClass: string;
-	role: string;
-	status: string;
-}
-
-export interface HomeTestimonial {
-	big?: boolean;
-	name: string;
-	quote: string;
-	role: string;
-}
-
-export interface ModerationAction {
-	color: Colors;
-	name: string;
-	temporary: Colors | null;
-	undo: Colors | null;
-}
-
-type LoggingEventDetailPart =
-	| { type: "mention"; name: string; avatar?: string }
-	| { type: "role"; name: string; color?: string }
-	| { type: "roles"; items: { name: string; color?: string }[] }
-	| { type: "text"; content: string };
-
-interface LoggingEventLogDetail {
-	label: string;
-	parts: LoggingEventDetailPart[];
-}
-
-export interface LoggingEventDetail {
-	tooltip: string;
-	title: string;
-	icon: string;
-	color: string;
-	action: string;
-	details: LoggingEventLogDetail[];
-}
 
 export const bentoFeatures: HomeFeature[] = [
 	{
@@ -528,32 +482,6 @@ export const Profiles = {
 	wolfstar: { app: true, name: "WolfStar", verified: true },
 } as const satisfies Record<string, Profile>;
 
-export interface Profile {
-	name: string;
-	app: boolean;
-	verified: boolean;
-}
-
-export type ProfileName = keyof typeof Profiles;
-
-/** Applications rendered inside the Discord slash-command suggestion panel. */
-export interface SlashCommandApp {
-	label: string;
-	/** Avatar image served from `public/avatars`. Takes precedence over `icon`. */
-	avatar?: string;
-	/** Icon fallback for apps without an avatar asset. */
-	icon?: string;
-}
-
-export type SlashCommandAppName =
-	| "catbot"
-	| "dyno"
-	| "fmbot"
-	| "utilsbot"
-	| "staryl"
-	| "wolfstar"
-	| "ring";
-
 export const SlashCommandApps: Record<SlashCommandAppName, SlashCommandApp> = {
 	catbot: { label: "Cat bot", icon: "ph:cat-fill" },
 	dyno: { label: "Dyno", icon: "ph:flame-fill" },
@@ -574,65 +502,6 @@ export const SlashCommandRailApps: SlashCommandAppName[] = [
 	"catbot",
 	"dyno",
 ];
-
-type ShowcaseCommandEmbedPart =
-	| { type: "mention"; name: string; avatar?: string }
-	| { type: "text"; content: string };
-
-interface ShowcaseCommandOption {
-	name: string;
-	value?: string;
-	description?: string;
-	focused?: boolean;
-}
-
-interface ShowcaseCommandEmbedLine {
-	label: string;
-	parts: ShowcaseCommandEmbedPart[];
-}
-
-interface ShowcaseCommandBase {
-	tooltip: string;
-	name: string;
-	subcommand?: string;
-	description: string;
-	invoker: ProfileName;
-	frequentlyUsed?: boolean;
-	options: ShowcaseCommandOption[];
-}
-
-interface ShowcaseCommandEmbedResponse {
-	responseType: "embed";
-	embedColor: string;
-	embedFooter: string;
-	embedLines: ShowcaseCommandEmbedLine[];
-}
-
-/** Plain success reply, e.g. "✅ Created case 3 | @baddie". */
-interface ShowcaseCommandTextResponse {
-	responseType: "text";
-	/** Prefix before the user mention; include trailing " | ". */
-	content: string;
-	mentionUser: string;
-	/** Desktop-only avatar inside the mention pill. */
-	mentionAvatar?: string;
-}
-
-interface ShowcaseCommandComponentsResponse {
-	responseType: "components";
-	accentColor: string;
-	lines: string[];
-	selectPlaceholder: string;
-	selectOptions: StringSelectMenuOption[];
-	buttonLabel: string;
-}
-
-export type ShowcaseCommand = ShowcaseCommandBase &
-	(
-		| ShowcaseCommandEmbedResponse
-		| ShowcaseCommandTextResponse
-		| ShowcaseCommandComponentsResponse
-	);
 
 export const showcaseCommands: ShowcaseCommand[] = [
 	{
@@ -827,13 +696,699 @@ export const showcaseCommands: ShowcaseCommand[] = [
 	},
 ];
 
-export interface OtherApp {
-	name: string;
-	explore: `/${string}`;
-	avatar: `/avatars/${string}`;
-	invite: string;
-	purposes: readonly string[];
-	description: string;
+export const EmojiRegexExtractName = /<?a?:(\w{2,32}):\d{17,21}>?/gi;
+
+/** Minimum vertical drag (px) before committing to the other snap. */
+export const DISCORD_APP_LAUNCHER_SHEET_DRAG_DISTANCE_PX = 56;
+
+/** Flick velocity (px/ms) that commits a snap even below the distance threshold. */
+export const DISCORD_APP_LAUNCHER_SHEET_DRAG_VELOCITY_PX_MS = 0.35;
+
+/**
+ * Preview row count for category cards on the main launcher scroll.
+ * “View More” appears only when a category has more entries than this.
+ */
+export const DISCORD_APP_LAUNCHER_CATEGORY_PREVIEW_COUNT = 4;
+
+/** Mobile “In This Server” / Recents tile strip length. */
+export const DISCORD_APP_LAUNCHER_MOBILE_TILE_COUNT = 2;
+
+/**
+ * Resolve the next sheet snap after a handle drag ends.
+ * Dragging down from `full` only collapses to `half` — it never dismisses the sheet.
+ */
+export function resolveDiscordAppLauncherSheetSnap(
+	options: ResolveDiscordAppLauncherSheetSnapOptions,
+): DiscordAppLauncherSheetSnap {
+	const { current, deltaY, velocityY } = options;
+	const distance = DISCORD_APP_LAUNCHER_SHEET_DRAG_DISTANCE_PX;
+	const velocity = DISCORD_APP_LAUNCHER_SHEET_DRAG_VELOCITY_PX_MS;
+
+	if (current === "half") {
+		if (deltaY <= -distance || velocityY <= -velocity) return "full";
+		return "half";
+	}
+
+	if (deltaY >= distance || velocityY >= velocity) return "half";
+	return "full";
 }
 
-export const EmojiRegexExtractName = /<?a?:(\w{2,32}):\d{17,21}>?/gi;
+/** Whether a category/list preview should show its overflow “View More” control. */
+export function shouldShowDiscordAppLauncherViewMore(
+	totalCount: number,
+	visibleCount: number,
+): boolean {
+	return totalCount > visibleCount;
+}
+
+/** Matches `.discord-string-select-menu-scroll` max-height — used for flip decisions. */
+export const STRING_SELECT_MENU_PANEL_MAX_HEIGHT = 190;
+
+/**
+ * Prefer opening above the trigger (Discord action-row / near-composer behavior).
+ * Flip below only when there is not enough room above and below has more space.
+ * Zero-sized rects (unlaid-out / jsdom) stay above so tests and SSR stay stable.
+ */
+export function resolveStringSelectMenuPlacement(
+	rect: Pick<DOMRectReadOnly, "top" | "bottom" | "height">,
+	viewportHeight: number,
+	panelMaxHeight = STRING_SELECT_MENU_PANEL_MAX_HEIGHT,
+): StringSelectMenuPlacement {
+	if (rect.height <= 0) return "above";
+	const spaceAbove = rect.top;
+	const spaceBelow = viewportHeight - rect.bottom;
+	if (spaceAbove >= panelMaxHeight) return "above";
+	if (spaceBelow >= panelMaxHeight && spaceBelow > spaceAbove) return "below";
+	return "above";
+}
+
+/** Default Recents row for the Discord App Launcher demo. */
+export const discordAppLauncherRecents: readonly DiscordAppLauncherEntry[] = [
+	{
+		id: "wolfstar-conf-menu",
+		name: "WolfStar Beta",
+		description: "> Customizable moderation and more!",
+		avatar: "/avatars/wolfstar.png",
+		kind: "command",
+		commandName: "conf",
+		tileTitle: "conf menu",
+		tileSubtitle: "WolfStar Beta",
+	},
+	{
+		id: "flamey-commands",
+		name: "Flamey",
+		description: "Flamey is a highly customizable, interactive all-in-one Discord bot.",
+		icon: "ph:flame-fill",
+		iconBg: "oklch(58% 0.2 250)",
+		tileTitle: "commands",
+		tileSubtitle: "Flamey",
+	},
+	{
+		id: "staryl",
+		name: "Staryl",
+		avatar: "/avatars/staryl.png",
+	},
+	{
+		id: "grid-a",
+		name: "Blips",
+		icon: "ph:circles-three-fill",
+		iconBg: "oklch(55% 0.16 250)",
+	},
+	{
+		id: "grid-b",
+		name: "Pulse",
+		icon: "ph:squares-four-fill",
+		iconBg: "oklch(52% 0.14 255)",
+	},
+	{
+		id: "eightball",
+		name: "8 Ball",
+		icon: "ph:number-circle-eight-fill",
+		iconBg: "oklch(32% 0.06 260)",
+	},
+	{
+		id: "diamond",
+		name: "Gem",
+		icon: "ph:diamond-fill",
+		iconBg: "oklch(58% 0.18 250)",
+	},
+	{
+		id: "wolfy",
+		name: "Wolfy",
+		icon: "ph:grid-four-fill",
+		iconBg: "oklch(68% 0.12 255)",
+	},
+];
+
+/** Apps listed under “Apps in this Server”. */
+export const discordAppLauncherServerApps: readonly DiscordAppLauncherEntry[] = [
+	{
+		id: "wolfstar",
+		name: "WolfStar Beta",
+		description: "> Customizable moderation and more!",
+		avatar: "/avatars/wolfstar.png",
+	},
+	{
+		id: "flamey",
+		name: "Flamey",
+		description: "Flamey is a highly customizable, interactive all-in-one Discord bot.",
+		icon: "ph:flame-fill",
+		iconBg: "oklch(58% 0.2 250)",
+	},
+	{
+		id: "staryl",
+		name: "Staryl",
+		avatar: "/avatars/staryl.png",
+	},
+	{
+		id: "wolfy",
+		name: "Wolfy",
+		icon: "ph:grid-four-fill",
+		iconBg: "oklch(68% 0.12 255)",
+	},
+];
+
+/** Promoted activity cards, rendered as local CSS artwork rather than remote assets. */
+export const discordAppLauncherPromoted: readonly DiscordAppLauncherPromo[] = [
+	{
+		id: "wordle",
+		title: "Wordle",
+		subtitle: "The New York Times Games",
+		description: "Play Wordle from New York Times Games.",
+		icon: "ph:grid-nine",
+		iconBg: "oklch(92% 0.01 95)",
+		variant: "wordle",
+	},
+	{
+		id: "magic-garden",
+		title: "Magic Garden",
+		description: "Garden with friends! Grow magical plants together.",
+		icon: "ph:flower-lotus-fill",
+		iconBg: "oklch(63% 0.2 345)",
+		variant: "garden",
+	},
+	{
+		id: "farm-merge-valley",
+		title: "Farm Merge Valley",
+		description: "Farm, merge, grow and expand your land.",
+		icon: "ph:plant-fill",
+		iconBg: "oklch(72% 0.16 145)",
+		variant: "farm",
+	},
+	{
+		id: "watch-together",
+		title: "Watch Together",
+		description: "Create and watch shared playlists with friends.",
+		icon: "ph:youtube-logo-fill",
+		iconBg: "oklch(94% 0.006 270)",
+		variant: "watch",
+	},
+];
+
+/** Full Recents destination shown by the first View More control. */
+export const discordAppLauncherRecentsList: readonly DiscordAppLauncherEntry[] = [
+	{
+		id: "farm-merge-valley",
+		name: "Farm Merge Valley",
+		description: "Farm, merge, grow and expand your land! Welcome to the world of farming.",
+		icon: "ph:plant-fill",
+		iconBg: "oklch(70% 0.16 145)",
+	},
+	{
+		id: "labscore",
+		name: "labsCore",
+		description: "As of January 1, 2026, labsCore is no longer available.",
+		icon: "ph:flask-fill",
+		iconBg: "oklch(32% 0.01 270)",
+	},
+	{
+		id: "poker",
+		name: "Poker Night",
+		description: "A Texas hold 'em style card game where you can prove your poker skills.",
+		icon: "ph:spade-fill",
+		iconBg: "oklch(52% 0.16 305)",
+	},
+	{
+		id: "putt",
+		name: "Putt Party",
+		description: "Perfect your putts in this pumped-up version of a mini golf game.",
+		icon: "ph:golf-fill",
+		iconBg: "oklch(58% 0.16 145)",
+	},
+	{
+		id: "rythm",
+		name: "Rythm",
+		description: "Listen to music together with your friends anywhere on Discord.",
+		icon: "ph:waveform-fill",
+		iconBg: "oklch(35% 0.08 255)",
+		promoted: true,
+	},
+	{
+		id: "skyra-beta",
+		name: "Skyra Beta",
+		description: "Beta - She/Her - Moderation management · Dashboard available.",
+		icon: "ph:shield-star-fill",
+		iconBg: "oklch(61% 0.16 235)",
+	},
+	{
+		id: "whiteboard",
+		name: "Whiteboard",
+		description: "Draw, upload images, write, and use GIFs to create virtually anything.",
+		icon: "ph:hand-pointing-fill",
+		iconBg: "oklch(93% 0.01 270)",
+	},
+	{
+		id: "bobble",
+		name: "Bobble League",
+		description: "In this arcade-style soccer game use Power Plays and formations.",
+		icon: "ph:soccer-ball-fill",
+		iconBg: "oklch(70% 0.09 80)",
+	},
+	{
+		id: "watch-together",
+		name: "Watch Together",
+		description: "Create and watch shared playlists of YouTube videos with your friends.",
+		icon: "ph:youtube-logo-fill",
+		iconBg: "oklch(94% 0.006 270)",
+	},
+	{
+		id: "activities",
+		name: "Activities",
+		description: "Bot for starting voice channel activities, made by advaith.",
+		icon: "ph:rocket-launch-fill",
+		iconBg: "oklch(32% 0.01 270)",
+	},
+	{
+		id: "assyst",
+		name: "Assyst",
+		description: "Image editing and meme bot with additional fun features.",
+		icon: "ph:image-square-fill",
+		iconBg: "oklch(38% 0.08 65)",
+	},
+	{
+		id: "astro",
+		name: "Astro",
+		description: "The most unique and complete Discord bot for temporary voice channels.",
+		icon: "ph:planet-fill",
+		iconBg: "oklch(38% 0.14 300)",
+	},
+	{
+		id: "dsc-bot",
+		name: "dsc.bot",
+		description: "Your place to discover Discord bots.",
+		icon: "ph:discord-logo-fill",
+		iconBg: "oklch(18% 0.005 270)",
+	},
+	{
+		id: "emoji-stealer",
+		name: "Emoji Stealer",
+		description: "Steal emojis, stickers, and reactions from other servers.",
+		icon: "ph:cat-fill",
+		iconBg: "oklch(94% 0.006 270)",
+	},
+	{
+		id: "green-bot",
+		name: "Green-bot",
+		description: "A simple and user friendly Discord music bot.",
+		icon: "ph:music-notes-fill",
+		iconBg: "oklch(60% 0.18 140)",
+	},
+	{
+		id: "tictactoe",
+		name: "TicTacToe",
+		description: "TicTacToe Bot recreates the popular game on Discord.",
+		icon: "ph:grid-four-fill",
+		iconBg: "oklch(82% 0.16 170)",
+	},
+	{
+		id: "wordle",
+		name: "Wordle",
+		description: "Play Wordle from New York Times Games. You have 6 chances to guess.",
+		icon: "ph:grid-nine",
+		iconBg: "oklch(92% 0.01 95)",
+	},
+];
+
+/** Secondary list opened via Chill Together → View More. */
+export const discordAppLauncherChillTogether: readonly DiscordAppLauncherEntry[] = [
+	{
+		id: "watch-together",
+		name: "Watch Together",
+		description: "Create and watch shared playlists of YouTube videos with your friends.",
+		icon: "ph:youtube-logo-fill",
+		iconBg: "oklch(94% 0.006 270)",
+	},
+	{
+		id: "lofi",
+		name: "Lofi",
+		description: "Vibe with your friends with lofi music and a curated ambience.",
+		icon: "ph:headphones-fill",
+		iconBg: "oklch(42% 0.14 295)",
+	},
+	{
+		id: "whiteboard",
+		name: "Whiteboard",
+		description: "Draw, upload images, write, and use GIFs to create virtually anything.",
+		icon: "ph:hand-pointing-fill",
+		iconBg: "oklch(94% 0.006 270)",
+	},
+	{
+		id: "karaoke-party-battle",
+		name: "Karaoke Party Battle",
+		description: "Karaoke Party Battle is the first karaoke Discord music game.",
+		icon: "ph:microphone-stage-fill",
+		iconBg: "oklch(56% 0.2 330)",
+	},
+	{
+		id: "rythm",
+		name: "Rythm",
+		description: "Create and listen to shared playlists of music.",
+		icon: "ph:music-notes-fill",
+		iconBg: "oklch(55% 0.2 300)",
+		promoted: true,
+	},
+	{
+		id: "sketch",
+		name: "Sketch Heads",
+		description: "Draw and guess with friends.",
+		icon: "ph:pencil-simple-fill",
+		iconBg: "oklch(65% 0.16 55)",
+	},
+	{
+		id: "checkers",
+		name: "Checkers In The Park",
+		description: "Play checkers against bots or friends.",
+		icon: "ph:checkerboard-fill",
+		iconBg: "oklch(55% 0.14 250)",
+	},
+	{
+		id: "land-io",
+		name: "Land-io",
+		description: "Claim as much land as you can!",
+		icon: "ph:map-trifold-fill",
+		iconBg: "oklch(60% 0.14 145)",
+	},
+	{
+		id: "putt",
+		name: "Putt Party",
+		description: "Compete in a mini-golf tournament!",
+		icon: "ph:golf-fill",
+		iconBg: "oklch(58% 0.16 145)",
+	},
+	{
+		id: "gartic",
+		name: "Gartic Phone",
+		description: "Draw and write with friends.",
+		icon: "ph:phone-fill",
+		iconBg: "oklch(62% 0.18 55)",
+	},
+	{
+		id: "know-what",
+		name: "Know What I Meme",
+		description: "Guess the meme with friends.",
+		icon: "ph:smiley-sticker-fill",
+		iconBg: "oklch(70% 0.16 85)",
+	},
+	{
+		id: "bobble",
+		name: "Bobble League",
+		description: "Compete in a fun bobble soccer game.",
+		icon: "ph:soccer-ball-fill",
+		iconBg: "oklch(55% 0.14 250)",
+	},
+	{
+		id: "ask-away",
+		name: "Ask Away",
+		description: "Get to know your friends better.",
+		icon: "ph:chat-teardrop-dots-fill",
+		iconBg: "oklch(58% 0.18 300)",
+	},
+	{
+		id: "color-together",
+		name: "Color Together",
+		description: "Draw and color with friends.",
+		icon: "ph:palette-fill",
+		iconBg: "oklch(65% 0.18 25)",
+	},
+	{
+		id: "nitro-control",
+		name: "Nitro Control",
+		description: "Control Nitro animations together.",
+		icon: "ph:lightning-fill",
+		iconBg: "oklch(55% 0.2 300)",
+		promoted: true,
+	},
+	{
+		id: "valorant",
+		name: "VALORANT - Sage's Quarters",
+		description: "Discover VALORANT'S new map, through the eyes of Sage.",
+		icon: "ph:crosshair-fill",
+		iconBg: "oklch(45% 0.18 25)",
+		promoted: true,
+	},
+	{
+		id: "masters-of-the-universe",
+		name: "Masters of the Universe",
+		description: "Eternia is calling you home.",
+		icon: "ph:sword-fill",
+		iconBg: "oklch(28% 0.04 260)",
+		promoted: true,
+	},
+];
+
+export const discordAppLauncherPuzzleGames: readonly DiscordAppLauncherEntry[] = [
+	{
+		id: "daily-sudoku",
+		name: "Daily Sudoku Together",
+		description: "Dive into Daily Sudoku Together on Discord—your ultimate Sudoku break.",
+		icon: "ph:number-square-one-fill",
+		iconBg: "oklch(92% 0.01 270)",
+	},
+	{
+		id: "daily-word-wheel",
+		name: "Daily Word Wheel",
+		description: "From the makers of Words with Friends, enjoy a daily word puzzle.",
+		icon: "ph:circles-four-fill",
+		iconBg: "oklch(83% 0.15 115)",
+	},
+	{
+		id: "blockbuster",
+		name: "BlockBuster",
+		description: "Welcome to BlockBuster: Adventures Puzzle! Immerse yourself in puzzles.",
+		icon: "ph:squares-four-fill",
+		iconBg: "oklch(46% 0.16 255)",
+	},
+	{
+		id: "hex-puzzle",
+		name: "Hex Puzzle Adventure",
+		description: "A hex stacking puzzle game. Use strategic thinking to solve the board.",
+		icon: "ph:hexagon-fill",
+		iconBg: "oklch(62% 0.2 40)",
+	},
+	{
+		id: "letter-league",
+		name: "Letter League",
+		description: "Use letter tiles from your rack to compete with friends.",
+		icon: "ph:squares-four-fill",
+		iconBg: "oklch(75% 0.15 305)",
+	},
+];
+
+export const discordAppLauncherCardGames: readonly DiscordAppLauncherEntry[] = [
+	discordAppLauncherRecentsList[2]!,
+	{
+		id: "blazing-eights",
+		name: "Blazing 8s",
+		description: "Be the first to zero cards by swapping hands and skipping players.",
+		icon: "ph:number-eight-fill",
+		iconBg: "oklch(72% 0.16 155)",
+	},
+	{
+		id: "blackjack",
+		name: "Blackjack",
+		description: "The first visual Blackjack game on Discord.",
+		icon: "ph:cards-three-fill",
+		iconBg: "oklch(35% 0.05 250)",
+	},
+	{
+		id: "erth-poker",
+		name: "ERTH Poker",
+		description: "The ultimate party poker game on Discord.",
+		icon: "ph:spade-fill",
+		iconBg: "oklch(48% 0.18 330)",
+	},
+];
+
+export const discordAppLauncherStrategyGames: readonly DiscordAppLauncherEntry[] = [
+	{
+		id: "roll20",
+		name: "Roll20",
+		description: "Bring your story to life on the powerful, easy-to-use virtual tabletop.",
+		icon: "ph:dice-five-fill",
+		iconBg: "oklch(54% 0.2 335)",
+	},
+	{
+		id: "chess-in-the-park",
+		name: "Chess In The Park",
+		description: "Unleash your inner Grandmaster by playing multiple games of Chess.",
+		icon: "ph:horse-fill",
+		iconBg: "oklch(76% 0.16 135)",
+	},
+	{
+		id: "battle-tabs",
+		name: "BattleTabs",
+		description: "Viking Ship Battles! BattleTabs is a turn-based PVP strategy game.",
+		icon: "ph:sword-fill",
+		iconBg: "oklch(68% 0.17 170)",
+	},
+	{
+		id: "stratego-online",
+		name: "Stratego Online",
+		description: "The beloved strategy game, known for its tactical depth.",
+		icon: "ph:shield-chevron-fill",
+		iconBg: "oklch(52% 0.19 25)",
+	},
+];
+
+export const discordAppLauncherSocialActivities: readonly DiscordAppLauncherEntry[] = [
+	{
+		id: "codenames",
+		name: "Codenames Online",
+		description: "A hit party game for 4+ players, designed by Vlaada Chvátil.",
+		icon: "ph:detective-fill",
+		iconBg: "oklch(53% 0.12 200)",
+	},
+	{
+		id: "goober-dash",
+		name: "Goober Dash",
+		description: "A 2D battle royale where Goobers run, jump and dash to victory.",
+		icon: "ph:person-simple-run-fill",
+		iconBg: "oklch(67% 0.2 345)",
+	},
+	{
+		id: "colonist",
+		name: "Colonist",
+		description: "A Settlers of Catan alternative with strategy game features.",
+		icon: "ph:hexagon-fill",
+		iconBg: "oklch(69% 0.16 190)",
+	},
+	{
+		id: "know-what-social",
+		name: "Know What I Meme",
+		description: "The feeling when memes and inside jokes become a game.",
+		icon: "ph:smiley-sticker-fill",
+		iconBg: "oklch(32% 0.04 250)",
+	},
+];
+
+export const discordAppLauncherActionGames: readonly DiscordAppLauncherEntry[] = [
+	{
+		id: "poxel",
+		name: "Poxel.io",
+		description: "The ultimate FPS action game. Play Poxel.io online with friends.",
+		icon: "ph:crosshair-simple-fill",
+		iconBg: "oklch(69% 0.17 190)",
+	},
+	{
+		id: "smash-karts",
+		name: "Smash Karts",
+		description: "Smash Karts is a MOKBA (Multiplayer Online Kart Battle Arena).",
+		icon: "ph:car-profile-fill",
+		iconBg: "oklch(70% 0.16 305)",
+	},
+	{
+		id: "krunker",
+		name: "Krunker Strike FRVR",
+		description: "Enter the Krunker Strike FRVR universe and compete.",
+		icon: "ph:target-fill",
+		iconBg: "oklch(61% 0.16 190)",
+	},
+	{
+		id: "kour",
+		name: "Kour.io",
+		description: "A multiplayer IO FPS game that plunges you into intense action.",
+		icon: "ph:person-simple-run-fill",
+		iconBg: "oklch(72% 0.16 105)",
+	},
+	{
+		id: "shellshock",
+		name: "Shell Shockers",
+		description: "First-person shooter with eggs. Join a match and crack some shells.",
+		icon: "ph:egg-crack-fill",
+		iconBg: "oklch(78% 0.14 85)",
+	},
+	{
+		id: "tanki",
+		name: "Tanki IO",
+		description: "Drive tanks, collect power-ups, and battle friends in arenas.",
+		icon: "ph:truck-fill",
+		iconBg: "oklch(55% 0.14 145)",
+	},
+];
+
+export const discordAppLauncherBotGames: readonly DiscordAppLauncherEntry[] = [
+	{
+		id: "owo",
+		name: "OwO",
+		description: "OwO What's this? Type 'owo help' for a list of commands!",
+		icon: "ph:cat-fill",
+		iconBg: "oklch(73% 0.15 210)",
+	},
+	{
+		id: "dank-memer",
+		name: "Dank Memer",
+		description: "Grind, fight, wager, gamble, fish, collect, and more.",
+		icon: "ph:smiley-melting-fill",
+		iconBg: "oklch(45% 0.14 145)",
+	},
+	{
+		id: "epic-rpg",
+		name: "EPIC RPG",
+		description: "An RPG bot! Start playing with rpg start or /start.",
+		icon: "ph:sword-fill",
+		iconBg: "oklch(76% 0.14 200)",
+	},
+	{
+		id: "virtual-fisher",
+		name: "VirtualFisher",
+		description: "A game you can play in your server focused on fishing.",
+		icon: "ph:fish-simple-fill",
+		iconBg: "oklch(92% 0.01 270)",
+	},
+];
+
+/** Main-scroll category previews, in the same order as Discord's launcher. */
+export const discordAppLauncherCategories: readonly DiscordAppLauncherListView[] = [
+	{ id: "puzzle-games", title: "Puzzle Games", entries: discordAppLauncherPuzzleGames },
+	{ id: "card-games", title: "Card Games", entries: discordAppLauncherCardGames },
+	{
+		id: "strategy-games",
+		title: "Strategy and Board Games",
+		entries: discordAppLauncherStrategyGames,
+	},
+	{
+		id: "chill-together",
+		title: "Chill Together",
+		entries: discordAppLauncherChillTogether,
+	},
+	{
+		id: "social-activities",
+		title: "Social Activities",
+		entries: discordAppLauncherSocialActivities,
+	},
+	{ id: "action-games", title: "Action Games", entries: discordAppLauncherActionGames },
+	{ id: "bot-games", title: "Bot Games", entries: discordAppLauncherBotGames },
+];
+
+export const discordAppLauncherListViews: readonly DiscordAppLauncherListView[] = [
+	{
+		id: "recents",
+		title: "Recents",
+		entries: discordAppLauncherRecentsList,
+	},
+	{
+		id: "server-apps",
+		title: "Apps in this Server",
+		entries: discordAppLauncherServerApps,
+	},
+	{
+		id: "chill-together",
+		title: "Chill Together",
+		entries: discordAppLauncherChillTogether,
+	},
+	...discordAppLauncherCategories.filter((view) => view.id !== "chill-together"),
+];
+
+/**
+ * Splits a promoted-card title into ~2 lines for stylized garden/farm artwork.
+ * First word on line 1; remaining words on line 2 (e.g. "Farm Merge Valley" →
+ * "Farm" / "Merge Valley"). Single-word titles stay on one line.
+ */
+export function splitDiscordAppLauncherPromoTitle(title: string): readonly string[] {
+	const words = title.trim().split(/\s+/).filter(Boolean);
+	if (words.length === 0) return [];
+	const first = words[0];
+	if (first === undefined) return [];
+	if (words.length === 1) return [first];
+	return [first, words.slice(1).join(" ")];
+}
